Introduce helper module for shared methods
