+++ p2panda-rs/src/api/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod errors;
+pub mod helpers;
 mod next_args;
 mod publish;
 pub mod validation;
